+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 1.11, 2024-04-11
 ----------------
 - scale mouse heatmap to screen height
+- add fullscreen view to keyboard heatmaps
 - omit programs with no events from applications list on heatmap page
 - make keyboard heatmap image path configurable
 - fix logging Numpad operator keys in Linux (issue #30)